[MODORDERS-1098] Started converting api unit tests
