--- src/test/resources/po_listed_print_serial_without_renewalDateAndInterval.json
@@ -1,168 +1,0 @@
-{
-  "approved": true,
-  "assignedTo": "ab18897b-0e40-4f31-896b-9c9adc979a88",
-  "manualPo": false,
-  "notes": [
-    "ABCDEFGHIJKLMNO",
-    "ABCDEFGHIJKLMNOPQRST",
-    "ABCDEFGHIJKLMNOPQRSTUV"
-  ],
-  "ongoing": {
-    "isSubscription": true,
-    "manualRenewal": true,
-    "reviewPeriod": 0
-  },
-  "orderType": "Ongoing",
-  "poNumber": "268758",
-  "reEncumber": false,
-  "vendor": "d0fb5aa0-cdf1-11e8-a8d5-f2801f1b9fd1",
-  "workflowStatus": "Open",
-  "compositePoLines": [
-    {
-      "acquisitionMethod": "306489dd-0053-49ee-a068-c316444a8f55",
-      "alerts": [
-        {
-          "alert": "Receipt overdue",
-          "id": "9a665b22-9fe5-4c95-b4ee-837a5433c95d"
-        }
-      ],
-      "cancellationRestriction": false,
-      "cancellationRestrictionNote": "ABCDEFGHIJKLMNOPQRSTUVW",
-      "claims": [
-        {
-          "claimed": false,
-          "grace": 30,
-          "sent": "2018-10-09T00:00:00.000Z"
-        }
-      ],
-      "collection": false,
-      "contributors": [
-        {
-          "contributor": "Ed Mashburn",
-          "contributorNameTypeId": "fbdd42a8-e47d-4694-b448-cc571d1b44c3"
-        }
-      ],
-      "cost": {
-        "currency": "USD",
-        "listUnitPrice": 24.99,
-        "listUnitPriceElectronic": 22.99,
-        "quantityElectronic": 1,
-        "quantityPhysical": 1
-      },
-      "description": "ABCDEFGH",
-      "details": {
-        "productIds": [
-          {
-            "productId": "9780764354113",
-            "productIdType": "8261054f-be78-422d-bd51-4ed9f33c3422"
-          }
-        ],
-        "receivingNote": "ABCDEFGHIJKL",
-        "subscriptionFrom": "2018-10-09T00:00:00.000Z",
-        "subscriptionInterval": 824,
-        "subscriptionTo": "2020-10-09T00:00:00.000Z"
-      },
-      "donor": "ABCDEFGHIJKLM",
-      "eresource": {
-        "accessProvider": "d0fb5aa0-cdf1-11e8-a8d5-f2801f1b9fd1",
-        "activated": false,
-        "activationDue": 0,
-        "createInventory": "None",
-        "materialType": "1a54b431-2e4f-452d-9cae-9cee66c9a892",
-        "trial": false,
-        "userLimit": 10
-      },
-      "fundDistribution": [
-        {
-          "code": "HIST",
-          "distributionType": "percentage",
-          "encumbrance": "eb506834-6c70-4239-8d1a-6414a5b08ac3",
-          "fundId": "fb7b70f1-b898-4924-a991-0e4b6312bb5f",
-          "value": 80
-        },
-        {
-          "code": "GENRL",
-          "distributionType": "percentage",
-          "encumbrance": "0466cb77-0344-43c6-85eb-0a64aa2934e5",
-          "fundId": "1b6d3338-186e-4e35-9e75-1b886b0da53e",
-          "value": 20.0
-        }
-      ],
-      "isPackage": false,
-      "locations": [
-        {
-          "locationId": "fcd64ce1-6995-48f0-840e-89ffa2288371",
-          "quantity": 2,
-          "quantityElectronic": 1,
-          "quantityPhysical": 1
-        }
-      ],
-      "orderFormat": "P/E Mix",
-      "paymentStatus": "Awaiting Payment",
-      "physical": {
-        "createInventory": "Instance, Holding, Item",
-        "materialSupplier": "73d14bc5-d131-48c6-b380-f8e62f63c8b6",
-        "materialType": "2fa93835-ea37-479d-b133-1d2a2279fcd8",
-        "receiptDue": "2018-10-10T00:00:00.000Z",
-        "volumes": [
-          "vol.1"
-        ]
-      },
-      "poLineDescription": "ABCDEFGHIJKLMNOPQRSTUVWXY",
-      "poLineNumber": "268758-03",
-      "publicationDate": "2017",
-      "publisher": "Schiffer Publishing",
-      "receiptDate": "2018-10-09T00:00:00.000Z",
-      "receiptStatus": "Awaiting Receipt",
-      "reportingCodes": [
-        {
-          "code": "CODE1",
-          "description": "ABCDEF",
-          "id": "5926dcd7-85f5-4504-8283-712595ebc38b"
-        },
-        {
-          "code": "CODE2",
-          "description": "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
-          "id": "fa316c04-8101-4e72-8aaf-01281bac718f"
-        },
-        {
-          "code": "CODE3",
-          "description": "ABCDE",
-          "id": "ea68b696-3125-4940-bf91-1d128323473e"
-        }
-      ],
-      "requester": "Leo Bulero",
-      "rush": true,
-      "selector": "ABCD",
-      "source": "EDI",
-      "tags": {
-        "tagList": [
-          "imPortant  _Tag",
-          "   ",
-          ""
-        ]
-      },
-      "titleOrPackage": "Kayak Fishing in the Northern Gulf Coast",
-      "vendorDetail": {
-        "instructions": "ABCDEFG",
-        "noteFromVendor": "ABCDEFGHIKJKLMNOP",
-        "referenceNumbers": [
-          {
-            "refNumber": "123456-78",
-            "refNumberType": "Vendor title number",
-            "vendorDetailsSource": "OrderLine"
-          }
-        ],
-        "vendorAccount": "8910-10"
-      },
-      "metadata": {
-        "createdByUserId": "ab18897b-0e40-4f31-896b-9c9adc979a88",
-        "createdDate": "2010-10-08T03:53:00.000"
-      }
-    }
-  ],
-  "metadata": {
-    "createdByUserId": "ab18897b-0e40-4f31-896b-9c9adc979a88",
-    "createdDate": "2010-10-08T03:53:00.000"
-  }
-}
